openapi_server/controllers/default_controller.py
import connexion
from typing import Dict
from typing import Tuple
from typing import Union

from openapi_server.models.contenido import Contenido  # noqa: E501
from openapi_server import util


def contenidos_id_get(id):  # noqa: E501
    """Obtener detalles de un contenido

    Obtiene la información completa de un contenido específico, incluyendo descripción, duración, género, y calificaciones promedio. # noqa: E501

    :param id: 
    :type id: int

    :rtype: Union[Contenido, Tuple[Contenido, int], Tuple[Contenido, int, Dict[str, str]]
    """
    return 'do some magic!'


def obtener_catalogo(genero=None, orden=None):  # noqa: E501
    """Obtener catálogo completo de contenidos

    Devuelve una lista completa de todos los contenidos disponibles en la plataforma, con opciones para filtrar por género y ordenar por fecha o popularidad. # noqa: E501

    :param genero: Filtrar los contenidos por género.
    :type genero: str
    :param orden: Ordenar los contenidos por popularidad o fecha de lanzamiento.
    :type orden: str

    :rtype: Union[Contenido, Tuple[Contenido, int], Tuple[Contenido, int, Dict[str, str]]
    """
    return 'do some magic!'


def reproducir_contenido(id):  # noqa: E501
    """Reproducir contenido en alta calidad

    Permite la reproducción del contenido seleccionado en alta calidad. # noqa: E501

    :param id: 
    :type id: int

    :rtype: Union[None, Tuple[None, int], Tuple[None, int, Dict[str, str]]
    """
    return 'do some magic!'
